+++ packages/amplify-codegen-e2e-core/src/categories/lambda-function.ts
@@ -1,7 +1,7 @@
 import { nspawn as spawn, ExecutionContext, KEY_DOWN_ARROW, getCLIPath, getProjectMeta, getBackendAmplifyMeta, invokeFunction } from '..';
 import { Lambda } from 'aws-sdk';
 import { singleSelect, multiSelect, moveUp, moveDown } from '../utils/selectors';
-import * as glob from 'glob';
+import { sync } from 'globby';
 import * as path from 'path';
 import _ from 'lodash';
 import { loadFeatureFlags } from '../utils/feature-flags';
@@ -467,7 +467,7 @@ const getRuntimeDisplayName = (runtime: FunctionRuntimes) => {
 
 export function validateNodeModulesDirRemoval(projRoot) {
   let functionDir = path.join(projRoot, 'amplify', '#current-cloud-backend', 'function');
-  const nodeModulesDirs = glob.sync('**/node_modules', {
+  const nodeModulesDirs = sync('**/node_modules', {
     cwd: functionDir,
     absolute: true,
   });